+++ packages/chameleon-loader/src/cml-compile/json-handle.js
@@ -13,8 +13,11 @@ module.exports = function(loaderContext, jsonObject, cmlType, componentDeps) {
         componentDeps.push(filePath);
         components[key] = refUrl;
       } else {
-        delete components[key];
-        cmlUtils.log.error(`can't find component:${refUrl} in ${loaderContext.resourcePath}`);
+        // plugin开头的小程序插件不做处理
+        if (components[key].indexOf('plugin://') !== 0) {
+          delete components[key];
+          cmlUtils.log.error(`can't find component:${refUrl} in ${loaderContext.resourcePath}`);
+        }
       }
     })
   }