+++ packages/chameleon-loader/src/loader.js
@@ -17,6 +17,7 @@ const cmlUtils = require('chameleon-tool-utils');
 const prehandle = require('./utils/prehandle.js');
 const loaderMethods = require('./loaderMethods');
 const miniAppScript = require('./miniapp-script.js');
+const loadIcon = require('./load-icon.js');
 let jsonObject = {};
 
 module.exports = function (content) {
@@ -253,6 +254,7 @@ module.exports = function (content) {
     })
 
     usingComponents = prepareParseUsingComponents(usingComponents);
+    
     //cml 编译出wxml模板
     if (type !== 'app') {
       let parseTemplate = parts.template && parts.template[0];
@@ -286,9 +288,13 @@ module.exports = function (content) {
     Object.keys(newJsonObj.usingComponents).forEach(key=>{
       newJsonObj.usingComponents[key] = cmlUtils.handleSpecialChar(newJsonObj.usingComponents[key])
     });
+    //处理tabbar中配置的icon路径
+    if(type == 'app'){
+      loadIcon.handleApptabbar(newJsonObj,filePath,cmlType)
+    }
     let jsonResult = JSON.stringify(newJsonObj, '', 4);
     self.emitFile(emitJsonPath, jsonResult);
-
+    
     //cml
     parts.styles.forEach(function (style, i) {
       //微信小程序在使用组件的时候 不支持属性选择器