+++ packages/chameleon-tool/commanders/extPlatform.js
@@ -1,0 +1,8 @@
+
+
+module.exports = function({type, media}) {
+  const utils = require('./utils.js');
+  cml.media = media;
+  cml.log.startBuilding();
+  utils.startReleaseOne(media, type);
+}