+++ packages/chameleon-tool/configs/component_export/getWebExportConfig_new.js
@@ -181,7 +181,6 @@ module.exports = function(options) {
     exportConfig.externals = externals;
   }
 
-
   return merge.smart(getWebCommonConfig(options), exportConfig);