+++ packages/chameleon-tool/configs/utils.js
@@ -63,6 +63,25 @@ exports.cssLoaders = function (options) {
 
   // generate loader string to be used with extract text plugin
   function generateLoaders(loader, loaderOptions) {
+
+    // 扩展流程
+    if (cml.config.get().extPlatform && ~Object.keys(cml.config.get().extPlatform).indexOf(options.type)) {
+      let extLoaders = [
+        {
+          loader: 'mvvm-style-loader'
+        },
+        getPostCssLoader('extend'),
+        {
+          loader: loader + '-loader',
+          options: Object.assign({}, loaderOptions, {
+            sourceMap: false
+          })
+        }
+      ]
+      addMediaLoader(extLoaders, options.type);
+      return extLoaders;
+    }
+
     var loaders = [cssLoader];
     let result = [];
 
@@ -192,7 +211,6 @@ exports.updateEntry = function (updateEntryConfig) {
           source = parts.template[0].content;
           options = analyzeTemplate(source, options)
         }
-
       }
     });
     let usedBuildInTagMap = options.usedBuildInTagMap;
@@ -491,7 +509,6 @@ let babelNpm = [
   'chameleon-api',
   'chameleon-tool-utils',
   'chameleon-css-loader',
-  'chameleon-loader',
   'chameleon-miniapp-target',
   'chameleon-mixins',
   'chameleon-template-parse',
@@ -505,7 +522,9 @@ let babelNpm = [
   'webpack-check-plugin',
   'webpack-liveload-middleware',
   'chameleon-weex-vue-loader',
-  'babel-plugin-chameleon-import'
+  'babel-plugin-chameleon-import',
+  /mvvm-interface-parser/,
+  /chameleon-loader/
 ];
 
 exports.getBabelPath = function () {
@@ -515,8 +534,12 @@ exports.getBabelPath = function () {
     path.join(cml.root, 'configs')
   ]
   babelNpm.forEach(item => {
-    babelPath.push(path.join(cml.projectRoot, 'node_modules', item))
-    babelPath.push(path.join(cml.root, 'node_modules', item))
+    if (typeof item === 'string') {
+      babelPath.push(path.join(cml.projectRoot, 'node_modules', item))
+      babelPath.push(path.join(cml.root, 'node_modules', item))
+    } else if (item instanceof RegExp) {
+      babelPath.push(item)
+    }
   })
   let configBabelPath = cml.config.get().babelPath || [];
   return configBabelPath.concat(babelPath);